Add additional gitee remote as fallback options for some repos
Some of the repositories can currently not be synced to the Eclipse
Oniro mirror due to various technical challenges (e.g. file is bigger
than 100MB in git history but not lfs, etc).

While we are working on sorting these out we need a way to fallback to
the original gitee repos for these exceptions.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>